Hice un custom hook para limpiar el código para el componente de Productos
