+++ src/components/Navbar.jsx
@@ -23,13 +23,6 @@ export const Navbar = () => {
 
                     <NavLink 
                         className="nav-item nav-link" 
-                        to="/clientes"
-                    >
-                        Clientes
-                    </NavLink>
-
-                    <NavLink 
-                        className="nav-item nav-link" 
                         to="/proveedores"
                     >
                         Proveedores
